refactor: PostDto 중복 변경
